database/requirements.txt
eve==0.7.4
flask-cors==3.0.3

# Pin the Flask-PyMongo version to avoid a bug in v2.0.0 related to an invalid URI scheme for the MongoDB server.
Flask-PyMongo==0.5.2
